requirements.txt
anthropic
pyautogui
python-dotenv
